--- a/stack.yaml
+++ b/stack.yaml
@@ -4,4 +4,4 @@ packages:
 - location: basement/
   extra-dep: true
 extra-deps:
-- gauge-0.1.1
+- gauge-0.1.3