mainly optimizing validation (benefiting `UArray Word8 -> String`) and `toList :: String -> [char]`, but probably also improving other benchs.

after:
```
types/String/FromUTF8/ascii/String       mean 238.6 ns  ( +- 4.627 ns  )
types/String/FromUTF8/mascii/String      mean 967.1 ns  ( +- 11.53 ns  )
types/String/FromUTF8/uni1/String        mean 1.199 μs  ( +- 98.72 ns  )
types/String/FromUTF8/uni2/String        mean 829.9 ns  ( +- 24.39 ns  )
types/String/Unpack/ascii/String         mean 3.937 μs  ( +- 104.9 ns  )
types/String/Unpack/mascii/String        mean 8.178 μs  ( +- 150.0 ns  )
types/String/Unpack/uni1/String          mean 3.000 μs  ( +- 123.1 ns  )
types/String/Unpack/uni2/String          mean 2.231 μs  ( +- 107.6 ns  )
```

before:

```
types/String/FromUTF8/ascii/String       mean 330.0 ns  ( +- 16.28 ns  )
types/String/FromUTF8/mascii/String      mean 1.134 μs  ( +- 25.78 ns  )
types/String/FromUTF8/uni1/String        mean 1.527 μs  ( +- 13.97 ns  )
types/String/FromUTF8/uni2/String        mean 1.105 μs  ( +- 33.75 ns  )
types/String/Unpack/ascii/String         mean 10.17 μs  ( +- 92.49 ns  )
types/String/Unpack/mascii/String        mean 19.46 μs  ( +- 136.7 ns  )
types/String/Unpack/uni1/String          mean 6.792 μs  ( +- 44.58 ns  )
types/String/Unpack/uni2/String          mean 5.127 μs  ( +- 205.1 ns  )
```