@vincenthz 2017-10-30
@ndmitchell should be of interest, should lower runtime/allocations in weeder