@vincenthz 2017-10-30
@ndmitchell do we have a volunteer for merging back the code just like for the Mutable algorithms ? :)