@ciderale 2017-10-31
@vincenthz By "merging back like for the Mutable algorithm", do you mean something along the line of #414 that I did with the sortByInplace? I could have look at further unifying more of the algorithms.