@vincenthz 2017-11-01
@ciderale yes, I think it would be interesting to see if we can do the same for the immutable bits.

I had a quick try when doing the reverse, but that turn out a bit painful to re-organise so I did the copy-pasta thing.