@ciderale 2017-11-01
I started working on the Algorithms in PrimArray. I'm about to prepare the PR for that. If that looks okay, I can continue on Strings.