@ndmitchell 2017-11-01
I benchmarked, but it's hard to spot a meaningful change in time (and like a fool, I deleted the old one before I ran memory). Useful to have anyway!